Eduardo

+++ Dto/Documents/DocumentCreatedDto.cs
@@ -4,8 +4,8 @@ public class DocumentCreatedDto
 { 
     public required string Title { get; set; }
     public required string Code { get; set; }
-    public required short Edition { get; set; }
-    public required short Pages { get; set; }
+    public required int Edition { get; set; }
+    public required int Pages { get; set; }
     public required DateTimeOffset DateOfValidity { get; set; }
     
     public required int TypeId { get; set; }
@@ -16,7 +16,7 @@ public class DocumentCreatedDto
     {
         return new Models.Document
         {
-            Code = Code, Edition = Edition, Pages = Pages, Title = Title,
+            Code = Code, Edition = (short)Edition, Pages = (short)Pages, Title = Title,
             DateOfValidity = DateOfValidity.UtcDateTime,
             TypeId = TypeId, ProcessId = ProcessId, ScopeId = ScopeId
         };
